src:arduino-timer: Convert for-loop macros to range-based for loops
Range-based loops (a C++11 feature) are available with the Arduino
compiler since IDE v1.6.6, released in Nov. 2015.  Use range-based loops
to keep reduced code duplication, but with C++ native code.

These range-based loops use references to the tasks array elements, so
the struct member access operator needs to be changed.

The remove() member function still expects a pointer to the task struct
to be removed from the tasks array.  So, changing from pointers to
references in the for-loop means using the address-of operator on the
task object to be removed.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>